refactor: user, common controller
